wip: refactoring of topics
