+++ src/components/content/taxonomy/topic-category.tsx
@@ -1,0 +1,69 @@
+import { faInfoCircle } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import clsx from 'clsx'
+
+import { useAuthentication } from '@/auth/use-authentication'
+import { Link } from '@/components/content/link'
+import { useInstanceData } from '@/contexts/instance-context'
+import { useLoggedInData } from '@/contexts/logged-in-data-context'
+import { TaxonomyLink, TopicCategoryTypes } from '@/data-types'
+import { categoryIconMapping } from '@/helper/icon-by-entity-type'
+
+export interface TopicCategoryProps {
+  links: TaxonomyLink[]
+  full?: boolean
+  category: TopicCategoryTypes
+  id?: number
+}
+
+export function TopicCategory({
+  links,
+  full,
+  category,
+  id,
+}: TopicCategoryProps) {
+  const { strings } = useInstanceData()
+  const loggedInData = useLoggedInData()
+  const auth = useAuthentication()
+
+  if (links.length === 0) return null
+
+  return (
+    <ul
+      key={category}
+      className={clsx('mt-5 first:mt-0', full && 'mb-6 mt-0 mobile:mt-2')}
+    >
+      <h4 className="text-truegray-900 text-lg mb-4 font-bold">
+        {strings.categories[category]}{' '}
+        <FontAwesomeIcon icon={categoryIconMapping[category]} />
+      </h4>
+
+      {links.map(renderLink)}
+    </ul>
+  )
+
+  function renderLink(link: TaxonomyLink, i: number) {
+    if (!auth.current && link.unrevised) return null
+    return (
+      <li className="block mb-3 leading-cozy" key={link.url + '_' + link.title}>
+        <Link
+          className={clsx(
+            'text-[1.2rem]',
+            link.unrevised ? 'opacity-60' : undefined
+          )}
+          href={link.url}
+          path={full ? [category, i] : [id!, category, i]}
+        >
+          {link.title}
+          {link.unrevised && (
+            <FontAwesomeIcon
+              icon={faInfoCircle}
+              title={loggedInData?.strings.revisions.unrevisedTaxNote}
+              className="ml-1"
+            />
+          )}
+        </Link>
+      </li>
+    )
+  }
+}
